Moved getAllowance back to lnbitsService
