+++ tabs/src/services/userService.ts
@@ -16,36 +16,6 @@ const getCurrentUser = async (): Promise<User | null> => {
   return user;
 };
 
-const getAllowance = async (
-  adminKey: string,
-  userId: string,
-): Promise<Allowance | null> => {
-  console.log(
-    `getNostrRewards starting ... (adminKey: ${adminKey}, stallId: ${userId})`,
-  );
-  try {
-    // TODO: Implement the actual API call to fetch the allowance
-    const allowance: Allowance = {
-      id: '123',
-      name: 'Allowance',
-      wallet: '123456789',
-      toWallet: '123456789',
-      amount: 25000,
-      startDate: new Date(),
-      endDate: null,
-      frequency: 'Monthly',
-      nextPaymentDate: new Date(new Date().setDate(new Date().getDate() + 7)),
-      lastPaymentDate: new Date(new Date().setDate(new Date().getDate() - 7)),
-      memo: "Don't spend it all at once",
-      active: true,
-    };
-    return allowance;
-  } catch (error) {
-    console.error(`Error fetching allowances for ${userId}:`, error);
-    throw error; // Re-throw the error to handle it in the parent function
-  }
-};
-
 const getAllUsers = async (): Promise<User[] | null> => {
   throw new Error('Not implemented');
 };
@@ -54,4 +24,4 @@ const getFriends = async (): Promise<User[] | null> => {
   throw new Error('Not implemented');
 };
 
-export { getCurrentUser, getAllUsers, getFriends, getAllowance, };
+export { getCurrentUser, getAllUsers, getFriends };
